Move imputation to utils to avoid circular import
